requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
google-generativeai>=0.3.0
googlemaps>=4.10.0
python-dateutil>=2.8.0